pyproject.toml
[tool.poetry]
name = "saa-revived"
version = "0.1.0"
description = "Steam accounts autoregistration. Automated 2fa binding"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "saa_revived"}]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.2"
bs4 = "^0.0.1"
cert-human = "^1.0.7"
rsa = "^4.9"
qiwipy = "^2.1.6"
asn1crypto = "^1.5.1"
pyopenssl = "^23.0.0"
sentry-sdk = "^1.16.0"
aiodns = "^3.0.0"
aiohttp = "^3.8.4"
aiosignal = "^1.3.1"
async-timeout = "^4.0.2"
attrs = "^22.2.0"
beautifulsoup4 = "^4.11.2"
certifi = "^2022.12.7"
cffi = "^1.15.1"
charset-normalizer = "^3.0.1"
cryptography = "^39.0.1"
distlib = "^0.3.6"
filelock = "^3.9.0"
2captcha-python = "^1.2.0"
frozenlist = "^1.3.3"
idna = "^3.4"
maxminddb = "^2.2.0"
multidict = "^6.0.4"
parse = "^1.19.0"
pipenv = "^2023.2.18"
platformdirs = "^3.0.0"
pyasn1 = "^0.4.8"
pycares = "^4.3.0"
pycparser = "^2.21"
python-dateutil = "^2.8.2"
six = "^1.16.0"
soupsieve = "^2.4"
urllib3 = "^1.26.14"
virtualenv = "^20.20.0"
virtualenv-clone = "^0.5.7"
yarl = "^1.8.2"
yapf = "^0.32.0"
capmonstercloudclient = "^1.1.0"
python-dotenv = "^1.0.0"
latest-user-agents = "^0.0.3"
pytest = "^7.2.2"
anticaptchaofficial = "^1.0.47"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
